Move Bool tests from Django backend to common
